bot_girls/requirements.txt
redis==3.5.3
python-telegram-bot==11.1.0
telegram==0.0.1       
google-api-python-client==2.37.0              
google-auth==2.6.0               
google-auth-httplib2==0.1.0               
google-auth-oauthlib==0.4.6               
# googleapis-common-protos==1.54.0